test: test BELLMAN_GPU_FRAMEWORK environment variable
Refactor the multicore environment variable test infrastructure into its
own module, so that it can also be used in other places.
