+++ src/gpu/program.rs
@@ -17,6 +17,38 @@ use crate::gpu::sources;
 use crate::gpu::GpuEngine;
 use pairing::Engine;
 
+/// Selects a CUDA or OpenCL on the `BELLMAN_GPU_FRAMEWORK` environment variable and the
+/// compile-time features.
+///
+/// You cannot select CUDA if the library was compiled without support for it.
+#[allow(clippy::unnecessary_wraps)] // No error can be returned if `cuda` and `opencl `are enabled.
+fn select_framework(default_framework: Framework) -> GPUResult<Framework> {
+    match env::var("BELLMAN_GPU_FRAMEWORK") {
+        Ok(env) => match env.as_ref() {
+            "cuda" => {
+                #[cfg(feature = "cuda")]
+                {
+                    Ok(Framework::Cuda)
+                }
+
+                #[cfg(not(feature = "cuda"))]
+                Err(GPUError::Simple("CUDA framework is not supported, please compile with the `cuda` feature enabled."))
+            }
+            "opencl" => {
+                #[cfg(feature = "opencl")]
+                {
+                    Ok(Framework::Opencl)
+                }
+
+                #[cfg(not(feature = "opencl"))]
+                Err(GPUError::Simple("OpenCL framework is not supported, please compile with the `opencl` feature enabled."))
+            }
+            _ => Ok(default_framework),
+        },
+        Err(_) => Ok(default_framework),
+    }
+}
+
 /// Returns the program for the preferred [`rust_gpu_tools::device::Framework`].
 ///
 /// If the device supports CUDA, then CUDA is used, else OpenCL. You can force a selection with
@@ -25,30 +57,7 @@ pub fn program<E>(device: &Device) -> GPUResult<Program>
 where
     E: Engine + GpuEngine,
 {
-    let framework = match env::var("BELLMAN_GPU_FRAMEWORK") {
-        Ok(env) => match env.as_ref() {
-            "cuda" => {
-                #[cfg(feature = "cuda")]
-                {
-                    Framework::Cuda
-                }
-
-                #[cfg(not(feature = "cuda"))]
-                return Err(GPUError::Simple("CUDA framework is not supported, please compile with the `cuda` feature enabled."));
-            }
-            "opencl" => {
-                #[cfg(feature = "opencl")]
-                {
-                    Framework::Opencl
-                }
-
-                #[cfg(not(feature = "opencl"))]
-                return Err(GPUError::Simple("OpenCL framework is not supported, please compile with the `opencl` feature enabled."));
-            }
-            _ => device.framework(),
-        },
-        Err(_) => device.framework(),
-    };
+    let framework = select_framework(device.framework())?;
     program_use_framework::<E>(device, &framework)
 }
 
@@ -75,6 +84,84 @@ where
                 .ok_or(GpuToolsError::DeviceNotFound)?;
             let program = opencl::Program::from_opencl(opencl_device, &src)?;
             Ok(Program::Opencl(program))
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::test_utils;
+
+    use super::*;
+
+    #[test]
+    fn test_bellman_gpu_framework() {
+        #[cfg(all(feature = "cuda", feature = "opencl"))]
+        {
+            // Use CUDA if set
+            test_utils::with_env_vars(vec![("BELLMAN_GPU_FRAMEWORK", Some("cuda"))], || {
+                assert_eq!(
+                    select_framework(Framework::Opencl).unwrap(),
+                    Framework::Cuda
+                );
+            });
+
+            // Use OpenCL if set
+            test_utils::with_env_vars(vec![("BELLMAN_GPU_FRAMEWORK", Some("opencl"))], || {
+                assert_eq!(
+                    select_framework(Framework::Cuda).unwrap(),
+                    Framework::Opencl
+                );
+            });
+
+            // Use default framework if arbitrary value is set
+            test_utils::with_env_vars(vec![("BELLMAN_GPU_FRAMEWORK", None)], || {
+                assert_eq!(select_framework(Framework::Cuda).unwrap(), Framework::Cuda);
+            });
+            test_utils::with_env_vars(vec![("BELLMAN_GPU_FRAMEWORK", None)], || {
+                assert_eq!(
+                    select_framework(Framework::Opencl).unwrap(),
+                    Framework::Opencl
+                );
+            });
+        }
+
+        #[cfg(feature = "cuda")]
+        {
+            // Use default value if no value is set
+            test_utils::with_env_vars(vec![("BELLMAN_GPU_FRAMEWORK", None)], || {
+                assert_eq!(select_framework(Framework::Cuda).unwrap(), Framework::Cuda);
+            });
+
+            // Use default framework if arbitrary value is set
+            test_utils::with_env_vars(
+                vec![("BELLMAN_GPU_FRAMEWORK", Some("notcudaoropencl"))],
+                || {
+                    assert_eq!(select_framework(Framework::Cuda).unwrap(), Framework::Cuda);
+                },
+            );
+        }
+
+        #[cfg(feature = "opencl")]
+        {
+            // Use default value if no value is set
+            test_utils::with_env_vars(vec![("BELLMAN_GPU_FRAMEWORK", None)], || {
+                assert_eq!(
+                    select_framework(Framework::Opencl).unwrap(),
+                    Framework::Opencl
+                );
+            });
+
+            // Use default framework if arbitrary value is set
+            test_utils::with_env_vars(
+                vec![("BELLMAN_GPU_FRAMEWORK", Some("notcudaoropencl"))],
+                || {
+                    assert_eq!(
+                        select_framework(Framework::Opencl).unwrap(),
+                        Framework::Opencl
+                    );
+                },
+            );
         }
     }
 }
