+++ src/multicore.rs
@@ -126,59 +126,10 @@ fn log2_floor(num: usize) -> u32 {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::env::{self, VarError};
-    use std::panic::{self, RefUnwindSafe, UnwindSafe};
-    use std::sync::Mutex;
+pub mod tests {
+    use crate::test_utils;
 
     use super::*;
-
-    lazy_static! {
-        static ref SERIAL_TEST: Mutex<()> = Default::default();
-    }
-
-    /// Sets environment variables to the given value for the duration of the closure.
-    /// Restores the previous values when the closure completes or panics, before unwinding the panic.
-    pub fn with_env_vars<F>(kvs: Vec<(&str, Option<&str>)>, closure: F)
-    where
-        F: Fn() + UnwindSafe + RefUnwindSafe,
-    {
-        let guard = SERIAL_TEST.lock().unwrap();
-        let mut old_kvs: Vec<(&str, Result<String, VarError>)> = Vec::new();
-        for (k, v) in kvs {
-            let old_v = env::var(k);
-            old_kvs.push((k, old_v));
-            match v {
-                None => env::remove_var(k),
-                Some(v) => env::set_var(k, v),
-            }
-        }
-
-        match panic::catch_unwind(|| {
-            closure();
-        }) {
-            Ok(_) => {
-                for (k, v) in old_kvs {
-                    reset_env(k, v);
-                }
-            }
-            Err(err) => {
-                for (k, v) in old_kvs {
-                    reset_env(k, v);
-                }
-                drop(guard);
-                panic::resume_unwind(err);
-            }
-        };
-    }
-
-    fn reset_env(k: &str, old: Result<String, VarError>) {
-        if let Ok(v) = old {
-            env::set_var(k, v);
-        } else {
-            env::remove_var(k);
-        }
-    }
 
     #[test]
     fn test_log2_floor() {
@@ -194,7 +145,7 @@ mod tests {
     #[test]
     fn test_read_num_cpus() {
         // use bellman if set
-        with_env_vars(
+        test_utils::with_env_vars(
             vec![("BELLMAN_NUM_CPUS", Some("6")), ("RAYON_NUM_THREADS", None)],
             || {
                 assert_eq!(read_num_cpus(), 6);
@@ -202,7 +153,7 @@ mod tests {
         );
 
         // bellman has priority over rayon
-        with_env_vars(
+        test_utils::with_env_vars(
             vec![
                 ("BELLMAN_NUM_CPUS", Some("6")),
                 ("RAYON_NUM_THREADS", Some("7")),
@@ -213,7 +164,7 @@ mod tests {
         );
 
         // use rayon if set, if bellman is not
-        with_env_vars(
+        test_utils::with_env_vars(
             vec![("BELLMAN_NUM_CPUS", None), ("RAYON_NUM_THREADS", Some("7"))],
             || {
                 assert_eq!(read_num_cpus(), 7);
@@ -221,7 +172,7 @@ mod tests {
         );
 
         // use num cpus if none is set
-        with_env_vars(
+        test_utils::with_env_vars(
             vec![("BELLMAN_NUM_CPUS", None), ("RAYON_NUM_THREADS", None)],
             || {
                 assert_eq!(read_num_cpus(), num_cpus::get());
